Share the email reader and its actions across pages and the relation manager.
