+++ packages/EmailIntegration/src/Filament/Pages/BaseRecordEmailsPage.php
@@ -9,18 +9,10 @@ use App\Models\Opportunity;
 use App\Models\People;
 use App\Models\User;
 use Filament\Actions\Action;
-use Filament\Forms\Components\Radio;
-use Filament\Forms\Components\Repeater;
-use Filament\Forms\Components\Select;
 use Filament\Notifications\Notification;
 use Filament\Resources\Pages\Concerns\InteractsWithRecord;
 use Filament\Resources\Pages\Page;
-use Filament\Schemas\Components\Grid;
-use Filament\Schemas\Components\Section;
-use Filament\Schemas\Components\Utilities\Get;
-use Filament\Support\Enums\Width;
 use Illuminate\Contracts\Database\Query\Builder;
-use Illuminate\Contracts\View\View;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphPivot;
 use Illuminate\Pagination\LengthAwarePaginator;
@@ -31,25 +23,21 @@ use Relaticle\EmailIntegration\Actions\ApproveEmailAccessRequestAction;
 use Relaticle\EmailIntegration\Actions\DenyEmailAccessRequestAction;
 use Relaticle\EmailIntegration\Actions\MarkAllEmailsAsReadAction;
 use Relaticle\EmailIntegration\Actions\MarkEmailAsReadAction;
-use Relaticle\EmailIntegration\Actions\RequestEmailAccessAction;
-use Relaticle\EmailIntegration\Actions\UpdateEmailSharingAction;
 use Relaticle\EmailIntegration\Enums\EmailAccessRequestStatus;
 use Relaticle\EmailIntegration\Enums\EmailFolder;
-use Relaticle\EmailIntegration\Enums\EmailPrivacyTier;
 use Relaticle\EmailIntegration\Filament\Concerns\HasEmailComposeActions;
 use Relaticle\EmailIntegration\Filament\Concerns\HasEmailFeatureFlag;
+use Relaticle\EmailIntegration\Filament\Concerns\HasEmailReaderActions;
 use Relaticle\EmailIntegration\Models\ConnectedAccount;
 use Relaticle\EmailIntegration\Models\Email;
 use Relaticle\EmailIntegration\Models\EmailAccessRequest;
-use Relaticle\EmailIntegration\Models\EmailShare;
-use Relaticle\EmailIntegration\Models\EmailThread;
 use Relaticle\EmailIntegration\Models\Scopes\VisibleEmailScope;
-use Relaticle\EmailIntegration\Services\EmailThreadSummaryService;
 
 abstract class BaseRecordEmailsPage extends Page
 {
     use HasEmailComposeActions;
     use HasEmailFeatureFlag;
+    use HasEmailReaderActions;
     use InteractsWithRecord;
     use WithPagination;
 
@@ -293,250 +281,6 @@ abstract class BaseRecordEmailsPage extends Page
         unset($this->emails);
     }
 
-    protected function manageSharingAction(): Action
-    {
-        return Action::make('manageSharing')
-            ->label(__('filament/pages/record-emails.actions.manage_sharing.label'))
-            ->icon('ri-share-line')
-            ->color('gray')
-            ->iconButton()
-            ->extraAttributes(['class' => 'fi-email-reader-action'])
-            ->tooltip(__('filament/pages/record-emails.actions.manage_sharing.label'))
-            ->modalHeading(__('filament/pages/record-emails.actions.manage_sharing.modal_heading'))
-            ->modalWidth(Width::FiveExtraLarge)
-            ->modalSubmitActionLabel(__('filament/pages/record-emails.actions.manage_sharing.submit'))
-            ->schema([
-                // Two halves of one decision, side by side: what the workspace at large
-                // gets, and who is named individually. Stacked, the second half reads as
-                // an afterthought below the fold.
-                Grid::make(['default' => 1, 'md' => 12])
-                    ->schema([
-                        // Cards rather than a select: each tier is a decision about who
-                        // sees what, which a dropdown of four nouns hides. Same cards as
-                        // the account settings page.
-                        Section::make(__('filament/pages/record-emails.fields.privacy_tier.label'))
-                            ->icon('heroicon-o-globe-alt')
-                            ->compact()
-                            ->columnSpan(['default' => 1, 'md' => 5])
-                            ->schema([
-                                Radio::make('privacy_tier')
-                                    ->hiddenLabel()
-                                    ->options(EmailPrivacyTier::class)
-                                    ->view('email-integration::forms.sharing-tier-cards')
-                                    ->viewData(['ariaLabel' => __('filament/pages/record-emails.fields.privacy_tier.label')])
-                                    ->required(),
-                            ]),
-
-                        Section::make(__('filament/pages/record-emails.fields.shares.label'))
-                            ->description(__('filament/pages/email-inbox.sharing.fields.shares.description'))
-                            ->icon('heroicon-o-user-group')
-                            ->compact()
-                            ->columnSpan(['default' => 1, 'md' => 7])
-                            ->schema([
-                                Repeater::make('shares')
-                                    ->hiddenLabel()
-                                    ->defaultItems(0)
-                                    // Who was added first says nothing about access, so
-                                    // the drag handle is just noise in the row header.
-                                    ->reorderable(false)
-                                    ->addActionLabel(__('filament/pages/email-inbox.sharing.fields.shares.add_action_label'))
-                                    // Not ->table(): that layout quietly stacks once its
-                                    // container is this narrow, which is what put the
-                                    // delete button on a line of its own. Naming the row
-                                    // instead gives each entry a header that carries the
-                                    // delete inline, and the two fields sit side by side.
-                                    ->itemLabel(fn (array $state): string => $this->shareRowLabel($state))
-                                    ->columns(2)
-                                    ->schema([
-                                        Select::make('tier')
-                                            ->label(__('filament/pages/record-emails.fields.tier.label'))
-                                            ->hiddenLabel()
-                                            ->options(EmailPrivacyTier::class)
-                                            ->disableOptionsWhenSelectedInSiblingRepeaterItems()
-                                            ->required(),
-
-                                        Select::make('shared_with')
-                                            ->label(__('filament/pages/record-emails.fields.shared_with.label'))
-                                            ->hiddenLabel()
-                                            ->placeholder(__('filament/pages/email-inbox.sharing.fields.shared_with.placeholder'))
-                                            ->options(fn (): array => $this->teammateOptions())
-                                            ->multiple()
-                                            ->searchable()
-                                            ->disableOptionsWhenSelectedInSiblingRepeaterItems()
-                                            ->required(),
-                                    ]),
-                            ]),
-                    ]),
-            ])
-            ->fillForm(function (array $arguments): array {
-                $email = $this->resolveTeamEmail($arguments['emailId'] ?? null, 'share');
-
-                if (! $email instanceof Email) {
-                    return [];
-                }
-
-                return [
-                    'privacy_tier' => $email->privacy_tier->value,
-                    'shares' => $this->shareFormRows($email),
-                ];
-            })
-            ->action(function (array $data, array $arguments): void {
-                $email = $this->resolveTeamEmail($arguments['emailId'] ?? null, 'share');
-
-                abort_if(! $email instanceof Email, 403);
-
-                resolve(UpdateEmailSharingAction::class)->execute(
-                    $email,
-                    $this->authUser(),
-                    $data['privacy_tier'] instanceof EmailPrivacyTier
-                        ? $data['privacy_tier']
-                        : EmailPrivacyTier::from($data['privacy_tier']),
-                    $data['shares'] ?? [],
-                );
-
-                Notification::make()
-                    ->success()
-                    ->title(__('filament/pages/record-emails.notifications.sharing_saved.title'))
-                    ->send();
-            });
-    }
-
-    /**
-     * @return array<int, array{tier: string, shared_with: array<int, int|string>}>
-     */
-    private function shareFormRows(Email $email): array
-    {
-        return $email->shares()
-            ->get()
-            ->groupBy(fn (EmailShare $share): string => $this->tierValue($share->tier))
-            ->map(fn (Collection $shares, string $tier): array => [
-                'tier' => $tier,
-                'shared_with' => $shares->pluck('shared_with')->all(),
-            ])
-            ->values()
-            ->all();
-    }
-
-    /**
-     * @param  array{tier?: mixed}  $state
-     */
-    private function shareRowLabel(array $state): string
-    {
-        $tier = $state['tier'] ?? null;
-
-        if ($tier instanceof EmailPrivacyTier) {
-            return $tier->getLabel();
-        }
-
-        if (is_string($tier) || is_int($tier)) {
-            return EmailPrivacyTier::tryFrom((string) $tier)?->getLabel()
-                ?? __('filament/pages/email-inbox.sharing.fields.shares.new_item');
-        }
-
-        return __('filament/pages/email-inbox.sharing.fields.shares.new_item');
-    }
-
-    private function tierValue(mixed $tier): string
-    {
-        if ($tier instanceof EmailPrivacyTier) {
-            return $tier->value;
-        }
-
-        if (is_string($tier) || is_int($tier)) {
-            return (string) $tier;
-        }
-
-        return '';
-    }
-
-    /**
-     * Teammates who can be given access to an email: everyone else on the team.
-     *
-     * @return array<string, string>
-     */
-    private function teammateOptions(): array
-    {
-        $user = $this->authUser();
-
-        return User::query()
-            ->where('current_team_id', $user->current_team_id)
-            ->where('id', '!=', $user->getKey())
-            ->orderBy('name')
-            ->pluck('name', 'id')
-            ->all();
-    }
-
-    protected function summarizeThreadAction(): Action
-    {
-        return Action::make('summarizeThread')
-            ->label(__('filament/pages/record-emails.actions.summarize_thread.label'))
-            ->icon('heroicon-o-sparkles')
-            ->color('gray')
-            ->iconButton()
-            ->extraAttributes(['class' => 'fi-email-reader-action'])
-            ->tooltip(__('filament/pages/record-emails.actions.summarize_thread.label'))
-            ->modalHeading(__('filament/pages/record-emails.actions.summarize_thread.modal_heading'))
-            ->modalIcon('heroicon-o-sparkles')
-            ->modalSubmitAction(false)
-            ->modalCancelActionLabel('Close')
-            ->modalContent(function (array $arguments): View {
-                $email = $this->resolveTeamEmail($arguments['emailId'] ?? null, 'viewBody');
-
-                if (! $email instanceof Email) {
-                    return view('filament.actions.ai-summary', ['summary' => null]);
-                }
-
-                return $this->buildThreadSummaryView($email);
-            });
-    }
-
-    protected function requestAccessAction(): Action
-    {
-        return Action::make('requestAccess')
-            ->label(__('filament/pages/record-emails.actions.request_access.label'))
-            ->icon('heroicon-o-key')
-            ->color('gray')
-            ->iconButton()
-            ->extraAttributes(['class' => 'fi-email-reader-action'])
-            ->tooltip(__('filament/pages/record-emails.actions.request_access.label'))
-            ->schema([
-                Select::make('tier_requested')
-                    ->label(__('filament/pages/record-emails.fields.tier_requested.label'))
-                    ->options([
-                        EmailPrivacyTier::SUBJECT->value => EmailPrivacyTier::SUBJECT->getLabel(),
-                        EmailPrivacyTier::FULL->value => EmailPrivacyTier::FULL->getLabel(),
-                    ])
-                    ->required(),
-            ])
-            ->action(function (array $data, array $arguments): void {
-                $email = $this->resolveTeamEmail($arguments['emailId'] ?? null, 'requestAccess');
-
-                abort_if(! $email instanceof Email, 403);
-
-                $request = resolve(RequestEmailAccessAction::class)->execute(
-                    $email,
-                    $this->authUser(),
-                    $data['tier_requested'] instanceof EmailPrivacyTier
-                        ? $data['tier_requested']
-                        : EmailPrivacyTier::from($data['tier_requested']),
-                );
-
-                if (! $request instanceof EmailAccessRequest) {
-                    Notification::make()
-                        ->warning()
-                        ->title(__('filament/pages/record-emails.notifications.pending_request.title'))
-                        ->send();
-
-                    return;
-                }
-
-                Notification::make()
-                    ->success()
-                    ->title(__('filament/pages/record-emails.notifications.access_request_sent.title'))
-                    ->send();
-            });
-    }
-
     protected function approveAccessRequestAction(): Action
     {
         return Action::make('approveAccessRequest')
@@ -603,63 +347,6 @@ abstract class BaseRecordEmailsPage extends Page
                     ->title(__('filament/pages/record-emails.notifications.access_request_denied.title'))
                     ->send();
             });
-    }
-
-    private function buildThreadSummaryView(Email $email): View
-    {
-        $thread = EmailThread::query()
-            ->where('thread_id', $email->thread_id)
-            ->where('connected_account_id', $email->connected_account_id)
-            ->first();
-
-        if ($thread === null) {
-            return view('filament.actions.ai-summary', ['summary' => null]);
-        }
-
-        $summary = resolve(EmailThreadSummaryService::class)
-            ->getSummary($thread, $this->authUser());
-
-        return view('filament.actions.ai-summary', ['summary' => $summary]);
-    }
-
-    /**
-     * Resolve an email by client-supplied id, scoped to the active team and gated by policy.
-     * Returns null when the email is outside the viewer's team or the ability is denied.
-     */
-    private function resolveTeamEmail(?string $emailId, string $ability): ?Email
-    {
-        if ($emailId === null) {
-            return null;
-        }
-
-        $user = $this->authUser();
-
-        $email = Email::query()
-            ->forTeam($user->current_team_id)
-            ->whereKey($emailId)
-            ->first();
-
-        if ($email === null) {
-            return null;
-        }
-
-        if (! $user->can($ability, $email)) {
-            return null;
-        }
-
-        return $email;
-    }
-
-    private function requesterNameForOwnedRequest(?string $requestId): string
-    {
-        if ($requestId === null) {
-            return 'this user';
-        }
-
-        return EmailAccessRequest::query()
-            ->whereKey($requestId)
-            ->where('owner_id', $this->authUser()->getKey())
-            ->first()?->requester->name ?? 'this user';
     }
 
     private function authUser(): User
