+++ packages/EmailIntegration/src/Filament/RelationManagers/BaseEmailsRelationManager.php
@@ -7,38 +7,29 @@ namespace Relaticle\EmailIntegration\Filament\RelationManagers;
 use App\Models\User;
 use Filament\Actions\Action;
 use Filament\Actions\ActionGroup;
-use Filament\Actions\ViewAction;
 use Filament\Forms\Components\Repeater;
 use Filament\Forms\Components\Select;
-use Filament\Infolists\Components\ViewEntry;
 use Filament\Notifications\Notification;
 use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Schemas\Schema;
-use Filament\Support\Enums\Width;
 use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
-use Illuminate\Contracts\View\View;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Arr;
-use Illuminate\Support\Collection;
-use Relaticle\EmailIntegration\Actions\RequestEmailAccessAction;
-use Relaticle\EmailIntegration\Actions\UpdateEmailSharingAction;
 use Relaticle\EmailIntegration\Enums\EmailDirection;
 use Relaticle\EmailIntegration\Enums\EmailPrivacyTier;
 use Relaticle\EmailIntegration\Filament\Actions\ConfigureMailboxAction;
 use Relaticle\EmailIntegration\Filament\Concerns\HasEmailComposeActions;
+use Relaticle\EmailIntegration\Filament\Concerns\HasEmailReaderActions;
 use Relaticle\EmailIntegration\Models\Email;
-use Relaticle\EmailIntegration\Models\EmailAccessRequest;
-use Relaticle\EmailIntegration\Models\EmailShare;
-use Relaticle\EmailIntegration\Models\EmailThread;
 use Relaticle\EmailIntegration\Models\Scopes\VisibleEmailScope;
 use Relaticle\EmailIntegration\Services\EmailSharingService;
-use Relaticle\EmailIntegration\Services\EmailThreadSummaryService;
 
 abstract class BaseEmailsRelationManager extends RelationManager
 {
     use HasEmailComposeActions;
+    use HasEmailReaderActions;
 
     protected static string $relationship = 'emails';
 
@@ -198,125 +189,12 @@ abstract class BaseEmailsRelationManager extends RelationManager
                     ->color('info'),
             ])
             ->recordActions([
-                ViewAction::make()
-                    ->modalHeading(__('filament/relation-managers/emails.actions.view.modal_heading'))
-                    ->modalWidth(Width::SevenExtraLarge)
-                    ->registerModalActions([
-                        $this->replyForwardEmailAction(),
-                    ])
-                    ->slideOver(),
+                $this->viewEmailAction(),
 
                 ActionGroup::make([
-                    Action::make('summarizeThread')
-                        ->label(__('filament/relation-managers/emails.actions.summarize_thread.label'))
-                        ->icon('heroicon-o-sparkles')
-                        ->color('gray')
-                        ->visible(fn (Email $record): bool => $record->user_id === $this->authUser()->getKey() || $this->authUser()->can('viewBody', $record))
-                        ->modalHeading(__('filament/relation-managers/emails.actions.summarize_thread.modal_heading'))
-                        ->modalIcon('heroicon-o-sparkles')
-                        ->modalSubmitAction(false)
-                        ->modalCancelActionLabel('Close')
-                        ->modalContent(fn (Email $record): View => $this->buildThreadSummaryView($record)),
-
-                    Action::make('manageSharing')
-                        ->label(__('filament/relation-managers/emails.actions.manage_sharing.label'))
-                        ->icon('heroicon-o-lock-open')
-                        ->modalHeading(__('filament/relation-managers/emails.actions.manage_sharing.modal_heading'))
-                        ->modalSubmitActionLabel('Save')
-                        ->visible(fn (Email $record): bool => $record->user_id === $this->authUser()->getKey())
-                        ->schema([
-                            Select::make('privacy_tier')
-                                ->label(__('filament/relation-managers/emails.fields.privacy_tier.label'))
-                                ->options(EmailPrivacyTier::class)
-                                ->required(),
-
-                            Repeater::make('shares')
-                                ->label(__('filament/relation-managers/emails.fields.shares.label'))
-                                ->defaultItems(0)
-                                ->addActionLabel('Add teammate')
-                                ->columns(2)
-                                ->schema([
-                                    Select::make('tier')
-                                        ->label(__('filament/relation-managers/emails.fields.tier.label'))
-                                        ->options(EmailPrivacyTier::class)
-                                        ->disableOptionsWhenSelectedInSiblingRepeaterItems()
-                                        ->required(),
-
-                                    Select::make('shared_with')
-                                        ->label(__('filament/relation-managers/emails.fields.shared_with.label'))
-                                        ->options(function (): array {
-                                            $user = $this->authUser();
-
-                                            return User::query()
-                                                ->where('current_team_id', $user->current_team_id)
-                                                ->where('id', '!=', $user->getKey())
-                                                ->pluck('name', 'id')
-                                                ->all();
-                                        })
-                                        ->multiple()
-                                        ->searchable()
-                                        ->disableOptionsWhenSelectedInSiblingRepeaterItems()
-                                        ->required(),
-                                ]),
-                        ])
-                        ->fillForm(fn (Email $record): array => [
-                            'privacy_tier' => $record->privacy_tier->value,
-                            'shares' => $this->shareFormRows($record),
-                        ])
-                        ->action(function (Email $record, array $data): void {
-                            abort_unless($this->authUser()->can('share', $record), 403);
-
-                            resolve(UpdateEmailSharingAction::class)->execute(
-                                $record,
-                                $this->authUser(),
-                                $data['privacy_tier'] instanceof EmailPrivacyTier
-                                    ? $data['privacy_tier']
-                                    : EmailPrivacyTier::from($data['privacy_tier']),
-                                $data['shares'] ?? [],
-                            );
-
-                            Notification::make()
-                                ->success()
-                                ->title(__('filament/relation-managers/emails.notifications.sharing_saved.title'))
-                                ->send();
-                        }),
-
-                    Action::make('requestAccess')
-                        ->label(__('filament/relation-managers/emails.actions.request_access.label'))
-                        ->icon('heroicon-o-key')
-                        ->visible(fn (Email $record): bool => $this->authUser()->cannot('viewBody', $record) && $this->authUser()->can('requestAccess', $record))
-                        ->schema([
-                            Select::make('tier_requested')
-                                ->label(__('filament/relation-managers/emails.fields.tier_requested.label'))
-                                ->options([
-                                    EmailPrivacyTier::SUBJECT->value => EmailPrivacyTier::SUBJECT->getLabel(),
-                                    EmailPrivacyTier::FULL->value => EmailPrivacyTier::FULL->getLabel(),
-                                ])
-                                ->required(),
-                        ])
-                        ->action(function (Email $record, array $data): void {
-                            $request = resolve(RequestEmailAccessAction::class)->execute(
-                                $record,
-                                $this->authUser(),
-                                $data['tier_requested'] instanceof EmailPrivacyTier
-                                    ? $data['tier_requested']
-                                    : EmailPrivacyTier::from($data['tier_requested']),
-                            );
-
-                            if (! $request instanceof EmailAccessRequest) {
-                                Notification::make()
-                                    ->warning()
-                                    ->title(__('filament/relation-managers/emails.notifications.pending_request.title'))
-                                    ->send();
-
-                                return;
-                            }
-
-                            Notification::make()
-                                ->success()
-                                ->title(__('filament/relation-managers/emails.notifications.access_request_sent.title'))
-                                ->send();
-                        }),
+                    $this->summarizeThreadAction(),
+                    $this->manageSharingAction(),
+                    $this->requestAccessAction(),
                 ]),
             ])
             ->emptyStateIcon('heroicon-o-envelope')
@@ -333,64 +211,12 @@ abstract class BaseEmailsRelationManager extends RelationManager
 
     public function infolist(Schema $schema): Schema
     {
-        return $schema
-            ->schema([
-                ViewEntry::make('email')
-                    ->hiddenLabel()
-                    ->view('filament.emails.email-view')
-                    ->columnSpanFull(),
-            ])
-            ->columns(1);
+        return $this->emailReaderInfolist($schema);
     }
 
     private function authUser(): User
     {
         /** @var User */
         return auth()->user();
-    }
-
-    /**
-     * @return array<int, array{tier: string, shared_with: array<int, int|string>}>
-     */
-    private function shareFormRows(Email $email): array
-    {
-        return $email->shares()
-            ->get()
-            ->groupBy(fn (EmailShare $share): string => $this->tierValue($share->tier))
-            ->map(fn (Collection $shares, string $tier): array => [
-                'tier' => $tier,
-                'shared_with' => $shares->pluck('shared_with')->all(),
-            ])
-            ->values()
-            ->all();
-    }
-
-    private function tierValue(mixed $tier): string
-    {
-        if ($tier instanceof EmailPrivacyTier) {
-            return $tier->value;
-        }
-
-        if (is_string($tier) || is_int($tier)) {
-            return (string) $tier;
-        }
-
-        return '';
-    }
-
-    private function buildThreadSummaryView(Email $email): View
-    {
-        $thread = EmailThread::query()->where('thread_id', $email->thread_id)
-            ->where('connected_account_id', $email->connected_account_id)
-            ->first();
-
-        if ($thread === null) {
-            return view('filament.actions.ai-summary', ['summary' => null]);
-        }
-
-        $summary = resolve(EmailThreadSummaryService::class)
-            ->getSummary($thread, $this->authUser());
-
-        return view('filament.actions.ai-summary', ['summary' => $summary]);
     }
 }
